refactor: refine dynamic string API names and improve test robustness
- Rename yio_append_s to yio_append_stream.
- Rename test_vreaprintf_growth.c to test_asprintf_growth.c.
- Add return value checks for yio_asprintf in parse and width tests.
- Remove redundant (size_t) cast in custom float formatter.
- Fix logic in yio_vasprintf and yio_vappend to correctly handle storage reuse and offsets.

+++ src/yio/private/yio_float_strfrom_custom.c
@@ -278,7 +278,7 @@ int YYIO_float_astrfrom_custom$1(YYIO_string *v, int precision0, char spec0, TYP
 		if (precision0 != 0) {
 			int exponent_tmp = 0;
 			(void)FREXP2(val, &exponent_tmp);
-			if (precision > (size_t)INT_MAX / 4) {
+			if (precision > INT_MAX / 4) {
 				return YIO_ERROR_ENOSYS;
 			}
 			const int bitpos = -5 + -4 * (int)precision + exponent_tmp;
